requirements.txt
# Profile stats generator and tests
pytest>=7.0.0
PyYAML>=6.0
